add table-header component and port 3001 to server
