+++ app/package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "bootstrap-daterangepicker": "^3.0.5",
+    "bootstrap-select": "^1.13.12",
     "concurrently": "^5.1.0",
     "jquery": "^3.4.1",
     "moment": "^2.24.0",
