+++ app/public/index.html
@@ -14,10 +14,13 @@
 
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-select@1.13.9/dist/css/bootstrap-select.min.css">
+
+
     <title>React App</title>
   </head>
   <body>
     <div id="root"></div>
-
+    
   </body>
 </html>
