--- app/src/components/date-picker/date-picker.js
@@ -1,16 +1,0 @@
-import React, { useState } from 'react';
-import DatePicker from 'react-datepicker';
-import "react-datepicker/dist/react-datepicker.css";
-import "./date-picker.scss"
-
-export default () => {
-  const [startDate, setStartDate] = useState(new Date());
-  return (
-    <DatePicker
-      dateFormat="yyyy/MM/dd"
-      selected={startDate}
-      onChange={date => setStartDate(date)}
-      className="customPicker"
-    />
-  );
-};
