--- app/src/components/date-picker/index.js
@@ -1,3 +1,0 @@
-import DatePicker from './date-picker';
-
-export default DatePicker;
