+++ app/src/components/pagination/pagination.js
@@ -4,7 +4,7 @@ import PaginationButton from "../pagination-button";
 const Pagination = ({ changeOffset, total }) => {
 
   
-  const countOfButtons = Math.floor(total/25) + 1
+  const countOfButtons = Math.round(total/25)
   let arr = [];
   for (let i = 0; i < countOfButtons; i++) {
     arr.push(<PaginationButton label={i} changeOffset={changeOffset} key={i}/>)
