+++ app/src/components/table-header/index.js
@@ -1,0 +1,3 @@
+import TableHeader from './table-header';
+
+export default TableHeader;
