+++ server.config.js
@@ -1,4 +1,4 @@
 module.exports = {
     host: 'localhost',
-    port: 3000
+    port: 3001
 };
